% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleadd.r
\name{demo1.add}
\alias{demo1.add}
\title{add function}
\usage{
demo1.add(a, b)
}
\arguments{
\item{a}{first integer}

\item{b}{second integer}
}
\value{
result
}
\description{
This function allows to add 2 integers
}
\examples{
demo1.add(10,20)
}
