% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addrange.r
\name{demo1.addrange}
\alias{demo1.addrange}
\title{add range function}
\usage{
demo1.addrange(from = 1, to = 10)
}
\arguments{
\item{from}{start integer}

\item{to}{second integer}
}
\value{
result
}
\description{
This function allows to add range of integers
}
\examples{
demo1.addrange(10,20)
}
