% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodbye.R
\name{goodbye}
\alias{goodbye}
\title{Goodbye function}
\usage{
goodbye(arg1)
}
\arguments{
\item{arg1}{this argument does nothing}
}
\value{
nothing
}
\description{
This function prints "goodbye" when it is called.
}
\examples{
goodbye("dummy arg")
}
